Return exceptions to caller
* Functions that can handle exceptions thrown by the file system API
will now return these to the caller. This allows the caller to provide
details of the error that is preventing the operation from succeeding to
the user.
* Modified function names to reflect new return types
* Fixed bug in folder deleting code that prevented it from deleting a
folder that had children.
* deleting a file will now return a boolean on success (False if the
file did not exists and True if it did) or an exception object if the
file could not be deleted and fbRetryOnFailure return False.
* writing data to a file will now return None on success or an exception
object on failure.

* Improved test coverage and resilience